+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH6W3Y80MP1AGJVQQDW38F7K"
+        "01KH84M7KYTTFVYBMFNERH75B2"
     )
     r"""The unique identifier of the data source"""