+++ packages/orq-rc/RELEASES.md
@@ -5559,3 +5559,13 @@ Based on:
 - [python v4.3.0-rc.56] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.56 - packages/orq-rc
+
+## 2026-02-12 05:21:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.711.0 (2.818.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.57] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.57 - packages/orq-rc